move pages test dedicated folder
