+++ tests/unit/pages/custom_pages/page.py
@@ -1,0 +1,5 @@
+import dash
+
+dash.register_page(__name__)
+
+layout = dash.html.Div("")
